Remove show_nr_indiv from model_dat and output
This is not really part of the data, IMO. If you want to know how many
results have been generated, you can just count them. If you want to
instruct a function to only generate a limited number of results, it is
better to tell that function than to pass around a limit inside your
data.