Use less strict versioning of dependencies
Since Causing is intended to be used as a library, dictating the exact
versions of all Causing-dependencies to the downstream project is not
desirable. To avoid too big changes in versioning, I specified the
versions using the compatible release specifier:
https://www.python.org/dev/peps/pep-0440/#compatible-release

There is no guarantee that all matching versions really are compatible,
but it's better than nothing.

Ideally, we would have a frozen list of specific versions to use in CI
in addition to the loose versions in setup.py, but I didn't add a
mechanism to handle that to avoid the complexity. If we ever change to
more modern tooling like poetry, we'll get that for free.